[VM]add swan option for isis (#7468)
What is the motivation for this PR?
Add one option for IS-IS test and once this option is enabled, all detected links will be enabled IS-IS protocol.

How did you do it?
If this option is enabled, all the detected links will be selected. If this option is not enabled, selected the first link of DUT.

How did you verify/test it?
Run IS-IS testcases, and check IS-IS neighbors.